+++ scripts/travis_prebuild.sh
@@ -17,7 +17,6 @@ else
   openssl aes-256-cbc -K "${encrypted_766accd2732a_key}" -iv "${encrypted_766accd2732a_iv}" -in "${SCRIPT_DIR}/id_rsa_gitlab-test.enc" -out ~/.ssh/id_rsa_gitlab-test -d
   cat >> ~/.ssh/config <<EOF
 Host gitlab.com
-  StrictHostKeyChecking no
   IdentityFile ~/.ssh/id_rsa_gitlab-test
 EOF
 fi